requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.24.0
lpips>=0.1.4
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
